Storybook uploads and start on worker service (#5)
- Add `/upload/storybook` API endpoint
- Create screenshot_tests row and upload to S3
- Create task queue
- `NOTIFY` on screenshot_tests row created or updated
- Stubbed out worker service

+++ cli/src/upload-storybook.ts
@@ -6,10 +6,29 @@ import { COMPRESSION_LEVEL, tar } from "zip-a-folder"
 
 import { info } from "./log"
 
+export type UploadStorybookOpts = {
+  storybookDir: string
+  commitSha: string
+  branchName: string
+  projectToken: string
+}
+
+type UploadResponse = { success: boolean; testId?: string; uploadId?: string; error?: string }
+
 /**
  * Packages a storybook build folder and uploads it to the vizdiff CDN.
  */
-export async function uploadStorybook(storybookDir: string, projectToken: string): Promise<void> {
+export async function uploadStorybook(opts: UploadStorybookOpts): Promise<void> {
+  const { storybookDir, commitSha, branchName, projectToken } = opts
+
+  // Sanity check input params
+  if (!isValidGitCommitHash(commitSha)) {
+    throw new Error(`Invalid commit SHA: "${commitSha}"`)
+  }
+  if (!branchName || branchName.length > 255) {
+    throw new Error(`Invalid branch name: "${branchName}"`)
+  }
+
   // Check that the storybook build manifest exists
   const projectJsonPath = path.join(storybookDir, "project.json")
   try {
@@ -33,7 +52,7 @@ export async function uploadStorybook(storybookDir: string, projectToken: string
   }
 
   // Create a tarball of the storybook build folder
-  const tarballFilename = path.join(os.tmpdir(), `${projectToken}.tar.gz`)
+  const tarballFilename = path.join(os.tmpdir(), `${randHexString()}.tar.gz`)
   const error = await tar(storybookDir, tarballFilename, { compression: COMPRESSION_LEVEL.medium })
   if (error) {
     throw new Error(`Failed to tar+gzip storybook build folder: ${error}`)
@@ -41,11 +60,20 @@ export async function uploadStorybook(storybookDir: string, projectToken: string
   info(`Created storybook tarball: ${tarballFilename}`)
 
   // POST the tarball to the vizdiff API
-  const url = `https://vizdiff.io/api/v1/upload-storybook?token=${projectToken}`
+  const body = await fs.readFile(tarballFilename)
+  const baseUrl = (process.env.VIZDIFF_API_URL ?? "https://vizdiff.io").replace(/\/+$/, "")
+  const url = `${baseUrl}/api/upload/storybook?token=${projectToken}`
+  info(
+    `Uploading ${formatBytes(
+      body.byteLength,
+    )} for commit ${commitSha} on branch ${branchName} to ${baseUrl}`,
+  )
   const response = await fetch(url, {
     method: "POST",
     headers: {
       "Content-Type": "application/gzip",
+      "X-Vizdiff-Commit-Sha": commitSha,
+      "X-Vizdiff-Branch-Name": branchName,
     },
     body: await fs.readFile(tarballFilename),
   })
@@ -54,8 +82,38 @@ export async function uploadStorybook(storybookDir: string, projectToken: string
       `Failed to upload storybook build folder to vizdiff CDN: ${response.statusText} (${response.status})`,
     )
   }
-  info(`Uploaded storybook build folder:`)
+  const res = (await response.json()) as UploadResponse
+  if (!res.success) {
+    throw new Error(
+      `Failed to upload storybook build folder to vizdiff CDN: ${JSON.stringify(res.error ?? res)}`,
+    )
+  }
+  info(`Uploaded storybook build folder, testId=${res.testId}, uploadId=${res.uploadId}`)
 
   // Delete the tarball
   await fs.unlink(tarballFilename)
+}
+
+function randHexString(length = 8): string {
+  return [...Array<undefined>(length)]
+    .map(() => Math.floor(Math.random() * 16).toString(16))
+    .join("")
+}
+
+function isValidGitCommitHash(hash: string): boolean {
+  const regex = /^[0-9a-f]{40}$/i
+  return regex.test(hash)
+}
+
+function formatBytes(bytes: number, decimals = 2): string {
+  if (bytes === 0) {
+    return "0 Bytes"
+  }
+
+  const k = 1024
+  const dm = decimals < 0 ? 0 : decimals
+  const sizes = ["Bytes", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"]
+  const i = Math.floor(Math.log(bytes) / Math.log(k))
+  const size = parseFloat((bytes / Math.pow(k, i)).toFixed(dm))
+  return `${size} ${sizes[i]}`
 }
